pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "archive-search"
version = "0.1.0"
description = "Mailing list archive search: index Google Groups + linked content, query via API and chat UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "trafilatura>=1.6",
    "pymupdf4llm>=0.0.12",
    "langchain-text-splitters>=0.2",
    "pymongo>=4.6",
    "openai>=1.12",
    "python-dotenv>=1.0",
    "click>=8.1",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "selenium>=4.15",
    "ollama>=0.3.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
]
local-embeddings = ["sentence-transformers>=2.3"]
ollama = ["ollama>=0.3.0"]

[project.scripts]
archive-indexer = "src.indexer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
